[ARROW-5461][Java] Move benchmarks to tests
